voice-gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
faster-whisper
piper-tts
numpy
python-multipart

